# feat: implement production-ready frame pool optimization (closes #42)